Created message handler
